project/requirements.txt
aiofiles==23.2.1
celery==5.4.0
fastapi==0.111.0
flower==2.0.1
pytest==8.2.2
redis==5.0.6
requests==2.32.3
uvicorn==0.30.1
httpx==0.27.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
python-dotenv==1.0.1
pydantic==2.7.4